# swagger.yaml
openapi: 3.0.3
info:
  title: Splitzies API
  description: |
    API for splitting receipts and assigning items to users.

    **Access model.** There are no accounts. `POST /devices` mints an anonymous
    device token, sent as `X-Device-Token`. Reading a bill needs only its ID,
    which is how a share link works. *Changing* a bill additionally requires
    that the calling device be on it — granted by uploading the receipt or by
    opening a live share link — so every mutating endpoint can answer `403
    not_a_participant`, and revoking a share link genuinely withdraws access.

    **Receipt photos are not stored.** The uploaded image is parsed in memory
    and discarded; only the extracted line items and receipt text are kept.

    **One device may keep 5 bills.** `POST /receipts/image` answers `409
    receipt_limit_reached` beyond that; deleting a bill frees a slot.
  version: 1.0.0
  contact:
    name: Splitzies
servers:
  - url: https://splitzies-backend-b74f565207ff.herokuapp.com
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
  /devices:
    post:
      summary: Create an anonymous device identity
      description: |
        Mints an anonymous identity so a user gets bill history without creating
        an account. The returned `device_token` is shown **once** — only its hash
        is stored — so the client must persist it (e.g. localStorage) and send it
        as the `X-Device-Token` header on subsequent requests. Losing the token
        loses that device's history.

        Every other endpoint treats the header as optional: a request without it
        behaves exactly as it did before devices existed.
      operationId: createDevice
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDeviceResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/receipts:
    get:
      summary: List this device's bill history
      description: |
        Returns bills this device created or joined, newest first. Summary rows
        only — fetch `GET /receipts/{receipt_id}` for items and assignments.
        `your_total` is present when the device has claimed a participant.
      operationId: listMyReceipts
      parameters:
        - $ref: '#/components/parameters/device_token'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: The `next_cursor` from the previous page (RFC3339).
      responses:
        '200':
          description: The device's bill history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMyReceiptsResponse'
        '400':
          description: Invalid limit or cursor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete this device's entire bill history
      description: |
        Empties this device's history. Every bill no other device is holding is
        deleted outright, along with its items, participants, assignments,
        share link and stored receipt text. Bills someone else still has are
        only removed from this device's history — a shared bill is not any one
        person's to destroy.

        The device identity itself is kept, so the app keeps working afterwards
        with nothing left attached to it.
      operationId: deleteAllMyReceipts
      parameters:
        - $ref: '#/components/parameters/device_token'
      responses:
        '200':
          description: History deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAllReceiptsResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/receipts/{receipt_id}:
    delete:
      summary: Delete a bill from this device's history
      description: |
        Removes the bill from this device's history and, when no other device
        is holding it, deletes the bill and everything on it outright. A bill
        someone else still has stays theirs — a shared bill is not any one
        person's to destroy.
      operationId: deleteMyReceipt
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - $ref: '#/components/parameters/device_token'
      responses:
        '200':
          description: Deleted, or removed from this device's history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt is not in this device's history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/share:
    post:
      summary: Create or fetch the receipt's share link
      description: |
        Returns a link that friends can open to edit the bill — no account
        required. Render `url` as a QR code and offer it as "copy link".

        Idempotent: repeat calls return the same active link, so re-opening the
        share sheet shows a stable QR code instead of invalidating one already
        sent. Anyone holding the link has full edit access to the bill.
      operationId: createShareLink
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - $ref: '#/components/parameters/device_token_optional'
      responses:
        '201':
          description: The receipt's active share link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Revoke the receipt's share link
      description: |
        Deactivates the link. Friends who already joined keep their access and
        the bill stays in their history — only the link stops working. A later
        POST mints a fresh link with a new token.
      operationId: revokeShareLink
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      responses:
        '200':
          description: Share link revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: The receipt has no active share link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /join/{token}:
    post:
      summary: Open a shared bill
      description: |
        Resolves a share token to its receipt and, when `X-Device-Token` is sent,
        adds the bill to that device's history (`joined: true`). Without a device
        token the caller can still edit the receipt — it just won't appear in any
        history. Use the returned `receipt_id` with the normal receipt endpoints.
      operationId: joinShareLink
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Share token from the share link URL
        - $ref: '#/components/parameters/device_token_optional'
      responses:
        '200':
          description: Share link resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinShareLinkResponse'
        '404':
          description: The share link is invalid or has been revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/users/{user_id}/claim:
    post:
      summary: Claim an existing participant as this device
      description: |
        Links an already-created participant to the calling device, for the case
        where the organiser typed everyone's names in before sharing the link.
        The claim shows up as `device_id` on that user in `GET /receipts/{id}`,
        which is how a client recognises which participant is itself. A device
        may claim at most one participant per receipt.

        To create *and* claim a participant in one call, post to
        `/receipts/{receipt_id}/users` with `claim: true` instead.
      operationId: claimReceiptUser
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: Receipt user ID to claim
        - $ref: '#/components/parameters/device_token'
      responses:
        '200':
          description: Participant claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Participant not found on this receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: This device already claims another participant on the receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/image:
    post:
      summary: Upload receipt image
      description: |
        Upload a receipt image to be parsed into items.

        The image is read in memory and **never stored** — it is sent to Gemini
        to be read, and discarded when the request ends. Only the parsed items
        and extracted receipt text are persisted.

        Requires a device token: the bill lands in that device's history, and
        the per-device limit of 5 bills is enforced here. This endpoint is the
        one that costs money per call and is rate limited more tightly than the
        rest of the API.
      operationId: uploadReceiptImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Receipt image file (JPEG, PNG, GIF, or WebP, max 10MB)
      responses:
        '201':
          description: Receipt uploaded and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadReceiptResponse'
        '400':
          description: Invalid request (missing image, unsupported file type, file too large)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: This device already holds the maximum of 5 bills (receipt_limit_reached)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}:
    get:
      summary: Get full receipt with bill split data
      description: |
        Returns the receipt with users, items, and assignments. Each assignment
        links a user to an item and includes amount_owed, computed as an equal
        split among all users assigned to that item (rounded to whole cents).

        **Polling for collaborators' edits.** This endpoint is conditional. Every
        response carries an `ETag` derived from a receipt version that every
        mutation bumps. Poll every few seconds sending the last `ETag` back as
        `If-None-Match`: an unchanged bill answers `304 Not Modified` after a
        single row read, and a changed one answers `200` with a new tag. Browser
        clients can read the header because `ETag` is in
        `Access-Control-Expose-Headers`.
      operationId: getReceipt
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
          description: ETag from the previous response, e.g. `"v7"`.
      responses:
        '200':
          description: Receipt with users, items, and assignments
          headers:
            ETag:
              schema:
                type: string
              description: Receipt version, e.g. `"v7"`. Send back as If-None-Match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetReceiptResponse'
        '304':
          description: The receipt has not changed since the supplied ETag
          headers:
            ETag:
              schema:
                type: string
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update receipt tax, tip, and/or extras_mode
      description: |
        Update tax, tip, and/or extras_mode on a receipt. Use tax/tip when
        values were not parsed from the receipt during upload. Only provided
        fields are updated.
      operationId: patchReceipt
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchReceiptRequest'
      responses:
        '200':
          description: Receipt updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "receipt updated successfully"
        '400':
          description: Invalid request (must include at least one of tax, tip, or extras_mode; extras_mode must be proportional or even)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/users:
    get:
      summary: Get users on a receipt
      description: List all users/participants associated with a receipt.
      operationId: getReceiptUsers
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetReceiptUsersResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add user to receipt
      description: |
        Add a participant to a receipt for bill splitting.

        Set `claim: true` (with an `X-Device-Token` header) to link the new
        participant to the calling device — this is what a friend does after
        opening a share link, and it makes the bill appear in their history.
        Leave it unset when typing in other people's names.
      operationId: addUserToReceipt
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - $ref: '#/components/parameters/device_token_optional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserToReceiptRequest'
      responses:
        '201':
          description: User added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddUserToReceiptResponse'
        '400':
          description: Invalid request (missing name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: claim was requested without a valid X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: This device already claims a participant on the receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/users/{user_id}:
    delete:
      summary: Remove a participant from a receipt
      description: |
        Removes a participant from the bill. Their item assignments are deleted
        with them, so anything only they had claimed becomes unclaimed again.
      operationId: removeUserFromReceipt
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: Receipt user ID to remove
      responses:
        '200':
          description: Participant removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Participant not found on this receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/items:
    get:
      summary: Get items on a receipt
      description: List all line items on a receipt.
      operationId: getReceiptItems
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      responses:
        '200':
          description: List of receipt items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetReceiptItemsResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add a unit to a receipt
      description: |
        Creates one individually-assignable unit. Pass group_id to add a unit
        to an existing group (the quantity stepper's "+"), or omit it to start
        a new group with this unit as its first member.
      operationId: createReceiptItem
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceiptItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptItem'
        '400':
          description: Invalid request (missing name/amount when group_id is not set)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt or group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    device_token:
      name: X-Device-Token
      in: header
      required: true
      schema:
        type: string
      description: Device token from POST /devices.
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "receipt \"01J...\" not found"
        code:
          type: string
          description: Machine-readable error code
          example: "receipt_not_found"
        request_id:
          type: string
          description: Request ID from X-Request-Id header (when provided)
          example: "f5542cdd-b900-260c-591c-c771cf2df890"
...
    CreateDeviceResponse:
      type: object
      properties:
        device_id:
          type: string
          description: Device ID (ULID). Matches device_id on a claimed participant.
          example: "01J8K2M3N4P5Q6R7S8T9V0W1X2"
        device_token:
          type: string
          description: |
            Opaque secret, returned once and never again — only its hash is
            stored. Persist it and send it as X-Device-Token; losing it loses
            this device's history.
...
    DeleteAllReceiptsResponse:
      type: object
      properties:
        message:
          type: string
          example: history deleted
        receipts_deleted:
          type: integer
          description: How many bills left this device's history
          example: 3
    ListMyReceiptsResponse:
      type: object
      properties:
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/ReceiptSummary'
        next_cursor:
          type: string
          format: date-time
          nullable: true
          description: Pass back as ?cursor= for the next page. Absent on the last page.